runfix(RichTextEditor): link editing when format buttons are not visible [WPB-16568] (#18872)
* runfix(RichTextEditor): link editting when format buttons are not visible

* refactor(FormatToolbar): create FormatButtons component
