+++ src/script/components/InputBar/InputBarEditor/RichTextEditor/FormatToolbar/FormatButtons/FormatButton/FormatButton.tsx
