+++ src/script/components/InputBar/InputBarEditor/RichTextEditor/FormatToolbar/FormatButtons/FormatButtons.styles.ts
@@ -1,0 +1,49 @@
+/*
+ * Wire
+ * Copyright (C) 2025 Wire Swiss GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see http://www.gnu.org/licenses/.
+ *
+ */
+
+import {CSSObject} from '@emotion/react';
+
+export const wrapperStyles: CSSObject = {
+  display: 'flex',
+  alignItems: 'center',
+  columnGap: '4px',
+  overflowX: 'auto',
+  scrollbarWidth: 'none',
+  '-ms-overflow-style': 'none',
+
+  '&::-webkit-scrollbar': {
+    width: 0,
+    height: 0,
+    display: 'none',
+  },
+};
+
+export const separatorStyles: CSSObject = {
+  width: '2px',
+  height: '24px',
+  backgroundColor: 'var(--gray-40)',
+
+  'body.theme-dark &': {
+    backgroundColor: 'var(--gray-80)',
+  },
+};
+
+export const separatorStylesEditing: CSSObject = {
+  backgroundColor: 'var(--accent-color-200)',
+};
