+++ src/script/components/InputBar/InputBarEditor/RichTextEditor/FormatToolbar/FormatButtons/common/isHeadingNode/isHeadingNode.ts
